Type of typeTag uint32_t -> uint8_t

+++ firmware/lib/ESPCamUART/ESPCamUART.h
@@ -18,7 +18,7 @@ public:
     ESPCamUART(HardwareSerial* camSerial, gpio_num_t pin) : camSerial{camSerial}, pin{pin} {};
     void startMeasurement();
     SensorValue getMeasurement();
-    uint32_t getTypeTag() const { return CAM_KEY; };
+    uint8_t getTypeTag() const { return CAM_KEY; };
 
     void updateNextSampleTime(uint32_t sampleInterval);
 };
